Improve clean up procedure of temporary files
The clean up was based on which files were not in the current folder, when
the compilation started. This failed completely, when the latex file is not
located in the current folder, e.g. sources/main.tex. Also when there had
been a run without `--clean` before, the temporary files were not new and
thus not removed.

This commit changes the procedure to fix these issues. Now a list of file
extensions of temporary LaTeX files is used. For each of them the project
name (latex path) without '.tex' extension is prepended. Every such file is
a temporary file that will be removed.